user Auth service added modified

+++ demofacebooked/src/main/java/com/facebooked/demofacebooked/SpringSecurity/service/UserAuthService.java
@@ -1,11 +1,13 @@
 package com.facebooked.demofacebooked.SpringSecurity.service;
 
 import com.facebooked.demofacebooked.SpringSecurity.model.UserAuth;
+import com.facebooked.demofacebooked.SpringSecurity.pojo.request.Role;
 import com.facebooked.demofacebooked.SpringSecurity.repo.UserAuthRepository;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -24,11 +26,18 @@ public class UserAuthService {
         userAuth.setPassword(passwordEncoder.encode(userAuth.getPassword()));
         return userAuthRepository.save(userAuth);
     }
-    @PreAuthorize(value = "USER")
+    //@PreAuthorize("USER")
     public UserAuth deleteUserAuth(Long id) {
-        System.out.println("delete UserAuth"+id);
-        Optional<UserAuth> userAuth1 = userAuthRepository.findById(id);
-        userAuthRepository.deleteById(id);
-        return userAuth1.orElseThrow();
+        System.out.println("delete UserAuth " + id);
+        Optional<UserAuth> userAuthOptional = userAuthRepository.findById(id);
+
+        if (userAuthOptional.isPresent()) {
+            UserAuth userAuth = userAuthOptional.get();
+            userAuthRepository.deleteById(id);
+            return userAuth;
+        } else {
+            throw new NoSuchElementException("UserAuth not found with id: " + id);
+        }
     }
+
 }
